# BE/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing seasons, rounds, and challenges in the Scout Arena application.",
        "title": "Scout Arena API",
        "contact": {
            "name": "API Support",
            "url": "http://www.scoutarena.com/support"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/seasons": {
            "post": {
                "description": "Tạo mùa giải mới với tên, hình ảnh và thời gian bắt đầu/kết thúc",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "seasons"
                ],
                "summary": "Tạo mới một mùa giải",
                "parameters": [
                    {
                        "description": "Thông tin mùa giải",
                        "name": "season",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SeasonRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.SeasonResponse"
                        }
                    },
                    "400": {
                        "description": "Lỗi validation",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/seasons/{id}": {
            "get": {
                "description": "Lấy thông tin chi tiết của một mùa giải dựa trên ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "seasons"
                ],
                "summary": "Lấy chi tiết mùa giải",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Season ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SeasonResponse"
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Cập nhật thông tin chi tiết của một mùa giải dựa trên ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "seasons"
                ],
                "summary": "Cập nhật mùa giải",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Season ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dữ liệu cập nhật",
                        "name": "season",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SeasonRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SeasonResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Xóa (soft delete) một mùa giải dựa trên ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "seasons"
                ],
                "summary": "Xóa mùa giải",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Season ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: Season deleted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.ArenaStatus": {
            "type": "string",
            "enum": [
                "DRAFT",
                "INCOMING",
                "ACTIVE",
                "FINISHED"
            ],
            "x-enum-varnames": [
                "ArenaDraft",
                "ArenaIncoming",
                "ArenaActive",
                "ArenaFinished"
            ]
        },
...
        "model.SeasonRequest": {
            "type": "object",
            "required": [
                "ended_at",
                "image_url",
                "name",
                "started_at",
                "status"
            ],
            "properties": {
                "ended_at": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.ArenaStatus"
                }
            }
        },
        "model.SeasonResponse": {
            "type": "object",
            "properties": {
                "ended_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                }
            }
        }
    }
}